added a getTarget() method
changed all this.target references to this.getTarget()

added a getParentUNID() method

changed the openForm method to now accept a config/options object instead of just a string for the formName.  You can now pass the following to the options object

options.formName (required)
options.dbPath (optional, defaults to this.dbPath)
options.isResponse (optional, defaults to false)
options.parentUNID (optional, defaults to this.document.universalUNID)

If you set isResponse to true then the call to /db/formName?OpenForm is changed to /db/formName?OpenForm&ParentUNID=unid

By making this change we can now support developers who need to compose response documents  

(we might later look into at support for response docs via the @Command([Compose];formName)  but that would require us knowing in advanced if the form was a response type or just a form that was set to accept values from it's parent form) -- but I seriously doubt it and believe at this time that developers can adjust.